Renamed org.drftpd.commands.nuke.common to org.drftpd.commands.nuke.metadata Moved keys that will be seriliazed into XML from main plugin to metadata plugin for request and stats. Closes ticket #321.

+++ src/plugins/org.drftpd.commands.nuke/src/org/drftpd/commands/nuke/NukeUtils.java
@@ -86,7 +86,7 @@ public class NukeUtils {
 						total = 0L;
 					}
 
-					total = total.longValue() + inode.getSize();
+					total = total + inode.getSize();
 					nukees.put(owner, total);
 				}
 			} catch (FileNotFoundException e) {
